refactor: rename server registry import paths and remove unused server registry file
