+++ server/bedrock/bedrock.service.ts
@@ -6,7 +6,7 @@ import {
   Message,
   SystemContentBlock,
 } from "@aws-sdk/client-bedrock-runtime";
-import { serverRegistry } from "../../tools/server/registry.js";
+import { serverRegistry } from "../../tools/registry.server.js";
 
 export interface BedrockServiceConfig {
   region: string;
