+++ tools/ldap-tool/server/ldap.types.ts
@@ -20,19 +20,19 @@ export interface LdapSearchResponse {
 }
 
 // Load and validate environment variables
-const requiredEnvVars = [
-  "LDAP_URL",
-  "LDAP_USER",
-  "LDAP_PASSWORD",
-  "LDAP_BASE",
-  "LDAP_DN",
-] as const;
+// const requiredEnvVars = [
+//   "LDAP_URL",
+//   "LDAP_USER",
+//   "LDAP_PASSWORD",
+//   "LDAP_BASE",
+//   "LDAP_DN",
+// ] as const;
 
-requiredEnvVars.forEach((key) => {
-  if (!process.env[key]) {
-    throw new Error(`Missing required environment variable: ${key}`);
-  }
-});
+// requiredEnvVars.forEach((key) => {
+//   if (!process.env[key]) {
+//     throw new Error(`Missing required environment variable: ${key}`);
+//   }
+// });
 
 export const ldapConfig: LdapConfig = {
   url: process.env.LDAP_URL!,
